moved spine class
