+++ src/matly/spines_class.py
@@ -1,0 +1,7 @@
+class SpineClass:
+    def __init__(self, label, visible, color):
+        self.spine = {
+            'visible': visible,
+            'color': color
+        }
+        self.label = label
